requirements.txt
Django==1.9.3
django-webpack-loader==0.2.4
python-twitter==3.5
djangorestframework==3.6.4
django-nose>=1.4.3
coverage>=4.5.2
mock>=2.0.0